Add new tabs with placeholders and prepare the folder structure
