+++ src/app/pages/PerpetualPage/components/CompetitionPage/components/Leaderboard/components/BestReturnTable/index.module.scss
